rename steady state files
