+++ stream/cost_model/steady_state_tensor_lifetime.py
@@ -6,9 +6,9 @@ import seaborn
 from matplotlib import patches
 from matplotlib import pyplot as plt
 
-from stream.workload.steady_state.steady_state_node import SteadyStateNode
-from stream.workload.steady_state.steady_state_tensor import SteadyStateTensor
-from stream.workload.steady_state.steady_state_workload import SteadyStateWorkload
+from stream.workload.steady_state.node import SteadyStateNode
+from stream.workload.steady_state.tensor import SteadyStateTensor
+from stream.workload.steady_state.workload import SteadyStateWorkload
 
 
 @dataclass(frozen=True)
