+++ stream/workload/steady_state/rolling_buffer.py
@@ -1,4 +1,4 @@
-from stream.workload.steady_state.steady_state_tensor import SteadyStateTensor
+from stream.workload.steady_state.tensor import SteadyStateTensor
 
 
 class SteadyStateRollingBuffer(SteadyStateTensor):
